move read_packets into os specific implementations
