+++ src-tauri/src/packets/read_packets_linux.rs
@@ -1,0 +1,8 @@
+use cargo::packets;
+
+pub async fn read_packets(
+    packet_sender: &tokio::sync::mpsc::Sender<(packets::opcodes::Pkt, Vec<u8>)>,
+    restart_receiver: &mut watch::Receiver<bool>,
+) {
+    
+}
